Circle
A class that represents a Circle. It has only one data property - its radius, and it is set through the constructor. The class needs to have getter and setter methods for its diameter - the setter needs to calculate the radius and change it and the getter needs to use the radius to calculate the diameter and return it.